requirements.txt
fastapi
uvicorn
sqlalchemy
cryptography
python-multipart
python-dotenv
openpyxl